+++ src/commands/_register.ts
@@ -1,5 +1,6 @@
 import type { BuiltApifyCommand } from '../lib/command-framework/apify-command.js';
 import { ActorIndexCommand } from './actor/_index.js';
+import { ActorCalculateMemoryCommand } from './actor/calculate-memory.js';
 import { ActorChargeCommand } from './actor/charge.js';
 import { ActorGetInputCommand } from './actor/get-input.js';
 import { ActorGetPublicUrlCommand } from './actor/get-public-url.js';
@@ -7,6 +8,7 @@ import { ActorGetValueCommand } from './actor/get-value.js';
 import { ActorPushDataCommand } from './actor/push-data.js';
 import { ActorSetValueCommand } from './actor/set-value.js';
 import { ActorsIndexCommand } from './actors/_index.js';
+import { AuthIndexCommand } from './auth/_index.js';
 import { BuildsIndexCommand } from './builds/_index.js';
 import { TopLevelCallCommand } from './call.js';
 import { InstallCommand } from './cli-management/install.js';
@@ -35,6 +37,7 @@ export const apifyCommands = [
 	// namespaces
 	ActorIndexCommand,
 	ActorsIndexCommand,
+	AuthIndexCommand,
 	BuildsIndexCommand,
 	DatasetsIndexCommand,
 	KeyValueStoresIndexCommand,
@@ -73,6 +76,7 @@ export const actorCommands = [
 	ActorGetPublicUrlCommand,
 	ActorGetInputCommand,
 	ActorChargeCommand,
+	ActorCalculateMemoryCommand,
 
 	// top-level
 	HelpCommand,